% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_df.R
\name{overlap_df}
\alias{overlap_df}
\title{Overlap two data frame containing coordinates}
\usage{
overlap_df(
  df1,
  df2,
  df1_chr_col = "chr",
  df1_start_col = "start",
  df1_end_col = "end",
  df1_0base = F,
  df2_chr_col = "chr",
  df2_start_col = "start",
  df2_end_col = "end",
  df2_0base = F,
  minoverlap = 0L
)
}
\arguments{
\item{df1}{data.frame 1, required}

\item{df2}{data.frame 2, required}

\item{df1_chr_col}{chr column name data.frame 1, default="chr"}

\item{df1_start_col}{start column name data.frame 1, default="start"}

\item{df1_end_col}{end column name data.frame 1, default="end"}

\item{df1_0base}{whether data.frame 1 start is 0-based, default=F}

\item{df2_chr_col}{chr column name data.frame 2, default="chr"}

\item{df2_start_col}{start column name data.frame 2, default="start"}

\item{df2_end_col}{end column name data.frame 2, default="end"}

\item{df2_0base}{whether data.frame 2 start is 0-based, default=F}

\item{minoverlap}{minimum overlap length, default=0L}
}
\value{
a list of data.frame, constituting overlap_df1 and overlap_df2
}
\description{
Overlap two data frame containing coordinates
}
\examples{
df1 = data.frame(chr=c("chr1","chr1"), start=c(2345,12345), end=c(12345,23456))
df2 = data.frame(chr=c("chr1","chr2"), start=c(2245,1245), end=c(2345,2456))
overlap_df(df1, df2)
}
