% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair2pair.R
\name{pair2pair}
\alias{pair2pair}
\title{pair-end coordinate data.frame intersect (pairToPair)}
\usage{
pair2pair(
  df1,
  df2,
  df1_chr1_col = "bait_chr",
  df1_start1_col = "bait_start",
  df1_end1_col = "bait_end",
  df1_chr2_col = "oe_chr",
  df1_start2_col = "oe_start",
  df1_end2_col = "oe_end",
  df1_0base = F,
  df2_chr1_col = "bait_chr",
  df2_start1_col = "bait_start",
  df2_end1_col = "bait_end",
  df2_chr2_col = "oe_chr",
  df2_start2_col = "oe_start",
  df2_end2_col = "oe_end",
  df2_0base = F,
  minoverlap = 0L
)
}
\arguments{
\item{df1}{data.frame 1, required. contains two coordinates}

\item{df2}{data.frame 2, required. contains two coordinates}

\item{df1_chr1_col}{data.frame 1 coordinate 1 chromosome column name, default = "bait_chr"}

\item{df1_start1_col}{data.frame 1 coordinate 1 start column name, default = "bait_start"}

\item{df1_end1_col}{data.frame 1 coordinate 1 end column name, default = "bait_end"}

\item{df1_chr2_col}{data.frame 1 coordinate 2 chromosome column name, default = "oe_chr"}

\item{df1_start2_col}{data.frame 1 coordinate 2 start column name, default = "oe_start"}

\item{df1_end2_col}{data.frame 1 coordinate 2 end column name, default = "oe_end"}

\item{df1_0base}{whether data.frame 1 start is 0-based, default = F}

\item{df2_chr1_col}{data.frame 2 coordinate 1 chromosome column name, default = "bait_chr"}

\item{df2_start1_col}{data.frame 2 coordinate 1 start column name, default = "bait_start"}

\item{df2_end1_col}{data.frame 2 coordinate 1 end column name, default = "bait_end"}

\item{df2_chr2_col}{data.frame 2 coordinate 2 chromosome column name, default = "oe_chr"}

\item{df2_start2_col}{data.frame 2 coordinate 2 start column name, default = "oe_start"}

\item{df2_end2_col}{data.frame 2 coordinate 2 start column name, default = "oe_start"}

\item{df2_0base}{whether data.frame 2 start is 0-based, default = F}

\item{minoverlap}{minimum overlap length, default=0L}
}
\value{
a list of data.frame, constituting overlap_df1 and overlap_df2
}
\description{
Both ends need to be overlaped with each other
}
\examples{
df1 = data.frame(
bait_chr=c("chr1","chr1"), bait_start=c(2345,12345), bait_end=c(12345,23456),
 oe_chr=c("chr1","chr1"), oe_start=c(2345,12345), oe_end=c(12345,23456)
 )
df2 = data.frame(
bait_chr=c("chr1","chr2"), bait_start=c(2245,1245), bait_end=c(2345,2456),
oe_chr=c("chr1","chr1"), oe_start=c(2345,12345), oe_end=c(12345,23456)
)
pair2pair(df1, df2)
}
